x86: Enable mmuinfo
