+++ src/arch/x86/mmu/mmu.c
@@ -11,6 +11,7 @@
 #include <kernel/panic.h>
 
 #include <asm/flags.h>
+#include <asm/cr_regs.h>
 
 #include <hal/mmu.h>
 #include <mem/vmem.h>
@@ -20,47 +21,6 @@
 static uintptr_t *ctx_table[0x100] __attribute__((aligned(MMU_PAGE_SIZE)));
 static int ctx_counter = 0;
 
-/*
- * Work with registers
- */
-
-static inline void set_cr3(unsigned int pagedir) {
-	asm ("mov %0, %%cr3": :"r" (pagedir));
-}
-
-static inline unsigned int get_cr3(void) {
-	unsigned int _cr3;
-	asm ("mov %%cr3, %0":"=r" (_cr3));
-	return _cr3;
-}
-
-static inline void set_cr0(unsigned int val) {
-	asm ("mov %0, %%cr0" : :"r" (val));
-}
-
-static inline unsigned int get_cr0(void) {
-	unsigned int _cr0;
-	asm ("mov %%cr0, %0":"=r" (_cr0));
-	return _cr0;
-}
-
-static inline void set_cr4(unsigned int val) {
-	asm ("mov %0, %%cr4" : :"r" (val));
-}
-
-static inline unsigned int get_cr4(void) {
-	unsigned int _cr4;
-	asm ("mov %%cr2, %0":"=r" (_cr4):);
-	return _cr4;
-}
-
-
-static inline unsigned int get_cr2(void) {
-	unsigned int _cr2;
-
-	asm ("mov %%cr2, %0":"=r" (_cr2):);
-	return _cr2;
-}
 
 /*
 #define PAGE_4MB         0x400000UL
@@ -200,34 +160,3 @@ uintptr_t mmu_pte_unpack(uintptr_t pte, int *flags) {
 
 	return pte & MMU_PAGE_MASK;
 }
-
-#if 0
-/* Page Table flags */
-
-void mmu_pte_set_writable(uintptr_t *pte, int val) {
-	if (val) {
-		*pte = *pte | MMU_PAGE_WRITABLE;
-	} else {
-		*pte = *pte & (~MMU_PAGE_WRITABLE);
-	}
-}
-
-void mmu_pte_set_usermode(uintptr_t *pte, int val) {
-	if (val) {
-		*pte = *pte | MMU_PAGE_USERMODE;
-	} else {
-		*pte = *pte & (~MMU_PAGE_USERMODE);
-	}
-}
-
-void mmu_pte_set_cacheable(uintptr_t *pte, int val) {
-	if (val & PROT_NOCACHE) {
-		*pte = *pte | MMU_PAGE_DISABLE_CACHE;
-	} else {
-		*pte = *pte & (~MMU_PAGE_DISABLE_CACHE);
-	}
-}
-
-void mmu_pte_set_executable(uintptr_t *pte, int val) {
-}
-#endif
